requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-Login
psycopg2-binary
python-dotenv
requests